Feat (favorite) : add FavoritePathService to delete(3)

+++ src/main/java/atdd/favorite/service/FavoritePathService.java
@@ -1,16 +1,16 @@
 package atdd.favorite.service;
 
-import atdd.favorite.application.dto.FavoritePathResponseView;
+import atdd.favorite.application.dto.FavoritePathListResponseView;
 import atdd.favorite.application.dto.FavoritePathRequestView;
+import atdd.favorite.application.dto.FavoritePathResponseView;
 import atdd.favorite.domain.FavoritePath;
 import atdd.favorite.domain.FavoritePathRepository;
 import atdd.path.application.GraphService;
 import atdd.path.domain.Station;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Service
@@ -25,8 +25,8 @@ public class FavoritePathService {
     }
 
     public FavoritePathResponseView create(FavoritePathRequestView requestView)
-        throws Exception {
-        if(requestView.getStartId() == requestView.getEndId()){
+            throws Exception {
+        if (requestView.getStartId() == requestView.getEndId()) {
             throw new IllegalArgumentException("출발역과 도착역이 같습니다.");
         }
         FavoritePath savedFavoritePath
@@ -39,9 +39,16 @@ public class FavoritePathService {
     public void delete(FavoritePathRequestView requestView) throws Exception {
         Optional<FavoritePath> favoritePathById
                 = favoritePathRepository.findById(requestView.getId());
-        if(favoritePathById.get().getEmail() != requestView.getEmail()){
+        if (Optional.empty().isPresent()) {
+            throw new NoSuchElementException("즐겨찾기에 등록되지 않은 지하철경로입니다.");
+        }
+        if (favoritePathById.get().getEmail() != requestView.getEmail()) {
             throw new IllegalArgumentException("즐겨찾기를 등록한 사람만 삭제할 수 있습니다.");
         }
         favoritePathRepository.delete(FavoritePath.of(requestView));
+    }
+
+    public FavoritePathListResponseView showAllFavoritePath(FavoritePathRequestView requestView) {
+        return null;
     }
 }
